金钱小数点取两位

+++ MM_FESCO/Module/Work/Approval/Model/ReimburseApprovalListModel.h
@@ -57,7 +57,7 @@
 
 @property (nonatomic, strong) NSString *empName;
 
-@property (nonatomic, assign) NSInteger moneySum;
+@property (nonatomic, assign) CGFloat moneySum;
 
 
 @end
