+++ MM_FESCO/Module/Work/Reimburse/Model/NOBookChooseModel.h
@@ -40,7 +40,7 @@
 
 @property (nonatomic, assign) NSInteger billNum;  // 发票数 /
 
-@property (nonatomic, assign) NSInteger moneyAmount;   // 金额  /
+@property (nonatomic, assign) CGFloat moneyAmount;   // 金额  /
 
 @property (nonatomic, strong) NSArray *picArray; // 图片信息数组
