+++ MM_FESCO/Module/Work/Reimburse/Model/ProgressReimburseModel.h
@@ -54,7 +54,7 @@
 
 @property (nonatomic, assign) NSInteger billNum;  // 发票数 /
 
-@property (nonatomic, assign) NSInteger moneyAmount;   // 金额  /
+@property (nonatomic, assign) CGFloat moneyAmount;   // 金额  /
 
 @property (nonatomic, strong) NSArray *pics;
